esm: moves all config logic to a dedicated file so we can remove the calls to require() and __dirname
